+++ src/main/java/com/codeit/HRBank/mapper/EmployeeMapper.java
@@ -6,8 +6,10 @@ import com.codeit.HRBank.dto.data.DepartmentDto;
 import com.codeit.HRBank.dto.data.EmployeeDto;
 import com.codeit.HRBank.dto.response.CursorPageResponseDepartmentDto;
 import com.codeit.HRBank.dto.response.CursorPageResponseEmployeeDto;
+
 import java.util.List;
 import java.util.stream.Collectors;
+
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Slice;
 import org.springframework.stereotype.Component;
@@ -24,28 +26,28 @@ public class EmployeeMapper {
         }
 
         return new EmployeeDto(
-                employee.getId(),
-                employee.getName(),
-                employee.getEmail(),
-                employee.getEmployeeNumber(),
-                employee.getDepartment().getId(),
-                employee.getDepartment().getName(),
-                employee.getPosition(),
-                employee.getHireDate(),
-                employee.getStatus(),
-                profileImageId
+            employee.getId(),
+            employee.getName(),
+            employee.getEmail(),
+            employee.getEmployeeNumber(),
+            employee.getDepartment().getId(),
+            employee.getDepartment().getName(),
+            employee.getPosition(),
+            employee.getHireDate(),
+            employee.getStatus(),
+            profileImageId
         );
     }
 
 
     // Slice<Employee>을 CursorPageResponseEmployeeDto<EmployeeDto>로 변환하는 메서드
     public CursorPageResponseEmployeeDto<EmployeeDto> toDtoSlice(
-            Slice<Employee> employeeSlice) {
+        Slice<Employee> employeeSlice) {
 
         // 1. Slice<Employee>의 content를 Stream을 이용해 List<EmployeeDto>로 변환
         List<EmployeeDto> dtoList = employeeSlice.getContent().stream()
-                .map(this::toDto)
-                .collect(Collectors.toList());
+            .map(this::toDto)
+            .collect(Collectors.toList());
 
         // 2. CursorPageResponseEmployeeDto에 필요한 커서 및 페이지 정보 계산
         Long nextIdAfter = null;
@@ -57,12 +59,12 @@ public class EmployeeMapper {
 
         // 3. 변환된 DTO 리스트와 커서 정보를 담아 최종 응답 객체 생성
         return new CursorPageResponseEmployeeDto<>(
-                dtoList,
-                null, // nextCursor는 필요시 추가
-                nextIdAfter,
-                employeeSlice.getSize(),
-                (int) employeeSlice.getNumberOfElements(),
-                employeeSlice.hasNext()
+            dtoList,
+            null, // nextCursor는 필요시 추가
+            nextIdAfter,
+            employeeSlice.getSize(),
+            employeeSlice.getNumberOfElements(),
+            employeeSlice.hasNext()
         );
     }
